move if_then to reification factory
